# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: DEAN Orchestration API
  description: |
    The DEAN (Distributed Evolutionary Agent Network) API provides comprehensive endpoints for managing
    evolutionary optimization of code through autonomous agents. This API enables creation and management
    of agents, evolution trials, pattern discovery, and system monitoring.
  version: 1.0.0
  contact:
    name: DEAN Support Team
    email: support@dean-orchestration.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.dean-orchestration.com
    description: Production server
  - url: https://staging.dean-orchestration.com
    description: Staging server
  - url: http://localhost:8082
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/agents:
    get:
      tags:
        - Agents
      summary: List all agents
      description: Retrieve a list of all agents with optional filtering
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [active, inactive, evolving]
          description: Filter by agent status
        - in: query
          name: language
          schema:
            type: string
            enum: [python, javascript, java]
          description: Filter by programming language
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
          description: Maximum number of results
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Number of results to skip
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'
                  total:
                    type: integer
                    example: 42
                  limit:
                    type: integer
                    example: 100
                  offset:
                    type: integer
                    example: 0
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Agents
      summary: Create new agent
      description: Create a new evolutionary agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
            examples:
              pythonAgent:
                value:
                  name: search-optimizer
                  language: python
                  capabilities:
                    - search
                    - optimize
                    - analyze
                  initial_fitness: 0.5
      responses:
        '201':
          description: Agent created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/agents/{agentId}:
    get:
      tags:
        - Agents
      summary: Get agent details
      description: Retrieve detailed information about a specific agent
      parameters:
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
          description: Agent ID
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags:
        - Agents
      summary: Update agent
      description: Update agent configuration
      parameters:
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgentRequest'
      responses:
        '200':
          description: Agent updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags:
        - Agents
      summary: Delete agent
      description: Remove an agent from the system
      parameters:
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Agent deleted successfully
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /api/patterns:
    get:
      tags:
        - Patterns
      summary: List discovered patterns
      description: Retrieve patterns discovered during evolution trials
      parameters:
        - in: query
          name: min_confidence
          schema:
            type: number
            format: float
            minimum: 0
            maximum: 1
            default: 0.7
          description: Minimum confidence score
        - in: query
          name: category
          schema:
            type: string
            enum: [optimization, refactoring, security, performance]
          description: Pattern category
        - in: query
          name: language
          schema:
            type: string
            enum: [python, javascript, java]
          description: Programming language
      responses:
        '200':
          description: List of patterns
          content:
            application/json:
              schema:
                type: object
                properties:
                  patterns:
                    type: array
                    items:
                      $ref: '#/components/schemas/Pattern'
                  total:
                    type: integer
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        name:
          type: string
          example: search-optimizer
        language:
          type: string
          enum: [python, javascript, java]
          example: python
        status:
          type: string
          enum: [active, inactive, evolving]
          example: active
        fitness_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          example: 0.85
        capabilities:
          type: array
          items:
            type: string
          example: [search, optimize, analyze]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        evolution_count:
          type: integer
          description: Number of evolution trials participated in
          example: 5
    CreateAgentRequest:
      type: object
      required:
        - name
        - language
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
          example: search-optimizer
        language:
          type: string
          enum: [python, javascript, java]
          example: python
        capabilities:
          type: array
          items:
            type: string
          example: [search, optimize, analyze]
        initial_fitness:
          type: number
          format: float
          minimum: 0
          maximum: 1
          default: 0.5
    UpdateAgentRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
        status:
          type: string
          enum: [active, inactive]
        capabilities:
          type: array
          items:
            type: string
...
    Pattern:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Async Handler Optimization
        category:
          type: string
          enum: [optimization, refactoring, security, performance]
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          example: 0.92
        impact_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          example: 0.85
        description:
          type: string
          example: Convert synchronous operations to async for improved performance
        language:
          type: string
          enum: [python, javascript, java]
        discovered_at:
          type: string
          format: date-time
        usage_count:
          type: integer
          description: Number of times pattern has been applied
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            code: AUTH_001
            details:
              message: Invalid or expired token
    BadRequestError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Bad Request
            code: VAL_001
            details:
              field: generations
              message: Value must be between 1 and 100
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Not Found
            code: NOT_001
            details:
              resource: agent
              id: 123e4567-e89b-12d3-a456-426614174000
...